@@ -86,40 +90,43 @@
     "@storybook/addon-links": "^3.3.12",
     "@storybook/react": "^3.3.12",
     "autoprefixer": "^8.0.0",
-    "babel-core": "^7.0.0-bridge.0",
+    "babel-core": "^7.0.0-0",
     "babel-eslint": "^8.2.1",
-    "babel-jest": "^22.2.0",
+    "babel-jest": "^22.4.3",
     "babel-plugin-dynamic-import-node": "^1.2.0",
     "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-optional-chaining": "^7.0.0-beta.3",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
     "bootstrap-loader": "^2.1.0",
     "bootstrap-sass": "^3.3.7",
     "case-sensitive-paths-webpack-plugin": "^2.1.1",
-    "css-loader": "^0.28.9",
-    "eslint": "^4.17.0",
+    "css-loader": "^0.28.11",
+    "eslint": "^4.19.0",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-loader": "^1.9.0",
+    "eslint-loader": "^2.0.0",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jest": "^21.7.0",
+    "eslint-plugin-jest": "^21.15.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-react": "^7.4.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "favicons-webpack-plugin": "0.0.7",
     "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^2.28.0",
+    "html-webpack-plugin": "^3.0.7",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^22.2.0",
     "json-loader": "^0.5.7",
     "node-sass": "^4.5.3",
-    "postcss-loader": "^2.0.6",
+    "postcss-loader": "^2.1.2",
     "pre-commit": "^1.2.2",
+    "raw-loader": "^0.5.1",
     "react-hot-loader": "^4.0.0-rc.0",
     "resolve-url-loader": "^2.0.3",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.20.1",
     "truffle-hdwallet-provider": "0.0.3",
-    "uglifyjs-webpack-plugin": "^1.1.4",
-    "url-loader": "^0.6.2",
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "url-loader": "^1.0.1",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.4"
   }
