@@ -2,31 +2,23 @@ import {
   initGnosisConnection,
   getCurrentBalance,
   getCurrentAccount,
-  calcMarketGasCost,
-  calcBuySharesGasCost,
-  calcSellSharesGasCost,
-  calcCategoricalEventGasCost,
-  calcScalarEventGasCost,
-  calcCentralizedOracleGasCost,
-  calcFundingGasCost,
-  calcRedeemWinningsGasCost,
   getGasPrice,
-  getEtherTokens,
+  getTokenSymbol,
+  getTokenBalance,
 } from 'api'
 
 import { timeoutCondition, getGnosisJsOptions } from 'utils/helpers'
-import { GAS_COST } from 'utils/constants'
-import { createAction } from 'redux-actions'
 import { findDefaultProvider } from 'integrations/store/selectors'
-
+import { createAction } from 'redux-actions'
 import { setActiveProvider } from 'integrations/store/actions'
+import { getTokenAddress } from 'utils/configuration'
 
 // TODO define reducer for GnosisStatus
 export const setGnosisInitialized = createAction('SET_GNOSIS_CONNECTION')
 export const setConnectionStatus = createAction('SET_CONNECTION_STATUS')
-export const setGasCost = createAction('SET_GAS_COST')
 export const setGasPrice = createAction('SET_GAS_PRICE')
-export const setEtherTokens = createAction('SET_ETHER_TOKENS')
+export const setTokenBalance = createAction('SET_TOKEN_BALANCE')
+export const setTokenSymbol = createAction('SET_TOKEN_NAME')
 
 export const NETWORK_TIMEOUT = process.env.NODE_ENV === 'production' ? 10000 : 2000
 
@@ -35,45 +27,14 @@ export const requestGasPrice = () => async (dispatch) => {
   dispatch(setGasPrice({ entityType: 'gasPrice', gasPrice }))
 }
 
-export const requestGasCost = (contractType, opts) => async (dispatch) => {
-  if (contractType === GAS_COST.MARKET_CREATION) {
-    calcMarketGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.BUY_SHARES) {
-    calcBuySharesGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.SELL_SHARES) {
-    calcSellSharesGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.CATEGORICAL_EVENT) {
-    calcCategoricalEventGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.SCALAR_EVENT) {
-    calcScalarEventGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.CENTRALIZED_ORACLE) {
-    calcCentralizedOracleGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.FUNDING) {
-    calcFundingGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.REDEEM_WINNINGS) {
-    calcRedeemWinningsGasCost(opts).then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  }
+export const requestTokenSymbol = tokenAddress => async (dispatch) => {
+  const tokenSymbol = await getTokenSymbol(tokenAddress)
+  dispatch(setTokenSymbol({ tokenAddress, tokenSymbol }))
 }
 
-export const requestEtherTokens = account => async (dispatch) => {
-  const etherTokens = await getEtherTokens(account)
-  dispatch(setEtherTokens({ entityType: 'etherTokens', account, etherTokens }))
+export const requestTokenBalance = (tokenAddress, accountAddress) => async (dispatch) => {
+  const tokenBalance = await getTokenBalance(tokenAddress, accountAddress)
+  dispatch(setTokenBalance({ tokenAddress, tokenBalance }))
 }
 
 /**
@@ -97,12 +58,12 @@ export const initGnosis = () => async (dispatch, getState) => {
         throw new Error('No account found')
       }
 
-      await dispatch(setGnosisInitialized({ initialized: true }))
-      await requestEtherTokens()
+      dispatch(setGnosisInitialized({ initialized: true }))
+      getTokenBalance(getTokenAddress(), await getCurrentAccount())
     }
   } catch (error) {
     console.warn(`Gnosis.js initialization Error: ${error}`)
-    await dispatch(setConnectionStatus({ connected: false }))
+    dispatch(setConnectionStatus({ connected: false }))
     return dispatch(setGnosisInitialized({ initialized: false, error }))
   }
 
