@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { NavLink } from 'react-router-dom'
 import autobind from 'autobind-decorator'
-import { collateralTokenToText } from 'components/CurrencyName'
+import CurrencyName from 'components/CurrencyName'
 import DecimalValue from 'components/DecimalValue'
 import { providerPropType } from 'utils/shapes'
 import { upperFirst } from 'lodash'
@@ -24,64 +24,94 @@ class Header extends Component {
 
   render() {
     const {
-      version, hasWallet, currentAccount, currentNetwork, currentBalance, currentProvider,
+      version,
+      hasWallet,
+      currentAccount,
+      currentNetwork,
+      currentBalance,
+      currentProvider,
+      isTournament,
+      logoPath,
+      smallLogoPath,
+      showScoreboard,
+      showGameGuide,
+      gameGuideType,
+      gameGuideURL,
+      tokenAddress,
     } = this.props
+
+    const logoVars = {}
+    if (isTournament) {
+      logoVars['--logoAnnotation'] = "'Powered by Gnosis'"
+      logoVars['--logoPath'] = `url("${logoPath}")`
+      logoVars['--smallLogoPath'] = `url("${smallLogoPath}")`
+    }
+
+    let gameGuideLink = <div />
+    if (showGameGuide) {
+      if (gameGuideType === 'default') {
+        gameGuideLink = (
+          <NavLink to="/game-guide" activeClassName={cx('active')} className={cx('navLink')}>
+            Game guide
+          </NavLink>
+        )
+      }
+
+      if (gameGuideType === 'link') {
+        gameGuideLink = (
+          <a href={gameGuideURL} className={cx('navLink')} target="_blank">
+            Game Guide
+          </a>
+        )
+      }
+    }
+
     return (
       <div className={cx('headerContainer')}>
         <div className={cx('container')}>
           <div className={cx('group', 'logo')}>
             <NavLink to="/markets/list">
-              <div className={cx('headerLogo', 'beta')} />
+              <div className={cx('headerLogo', 'beta')} style={logoVars} />
             </NavLink>
           </div>
-          <div className={cx('group', 'left', 'version')}>
-            {version}
-          </div>
+          <div className={cx('group', 'left', 'version')}>{version}</div>
           <div className={cx('group', 'left', 'navLinks')}>
             {hasWallet && (
-              <NavLink
-                to="/dashboard"
-                activeClassName={cx('navLink', 'active')}
-                className={cx('navLink')}
-              >
+              <NavLink to="/dashboard" activeClassName={cx('active')} className={cx('navLink')}>
                 Dashboard
               </NavLink>
             )}
-            <NavLink
-              to="/markets/list"
-              activeClassName={cx('navLink', 'active')}
-              className={cx('navLink')}
-            >
+            <NavLink to="/markets/list" activeClassName={cx('active')} className={cx('navLink')}>
               Markets
             </NavLink>
             {hasWallet && (
-              <NavLink
-                to="/transactions"
-                activeClassName={cx('navLink', 'active')}
-                className={cx('navLink')}
-              >
+              <NavLink to="/transactions" activeClassName={cx('active')} className={cx('navLink')}>
                 Transactions
               </NavLink>
             )}
+            {showScoreboard && (
+              <NavLink to="/scoreboard" activeClassName={cx('active')} className={cx('navLink')}>
+                Scoreboard
+              </NavLink>
+            )}
+            {gameGuideLink}
           </div>
 
           <div className={cx('group', 'right')}>
             {hasWallet &&
               currentProvider && (
-                <div className={cx('account')}>
-                  {currentNetwork &&
+              <div className={cx('account')}>
+                {currentNetwork &&
                     currentNetwork !== 'MAIN' && (
-                      <span className={cx('network', 'text')}>
-                        Network: {upperFirst(currentNetwork.toLowerCase())}
-                      </span>
-                    )}
-                  <DecimalValue value={currentBalance} className={cx('balance', 'test')} />&nbsp;
-                  <span className={cx('account', 'text')}>{collateralTokenToText()}</span>
-                  <ProviderIcon provider={currentProvider} />
-                  <Identicon account={currentAccount} />
-                  <MenuAccountDropdown />
-                </div>
-              )}
+                    <span className={cx('network', 'text')}>Network: {upperFirst(currentNetwork.toLowerCase())}</span>
+                )}
+                <DecimalValue value={currentBalance} className={cx('balance', 'test')} />&nbsp;
+                <CurrencyName className={cx('account', 'text')} tokenAddress={tokenAddress} />
+                <ProviderIcon provider={currentProvider} />
+                <Identicon account={currentAccount} />
+                <MenuAccountDropdown />
+              </div>
+            )}
             {!hasWallet && (
               <a className={cx('connect-wallet')} onClick={this.handleConnectWalletClick}>
                 Connect a wallet
@@ -102,6 +132,14 @@ Header.propTypes = {
   currentProvider: providerPropType,
   currentAccount: PropTypes.string,
   openConnectWalletModal: PropTypes.func.isRequired,
+  isTournament: PropTypes.bool,
+  logoPath: PropTypes.string.isRequired,
+  smallLogoPath: PropTypes.string.isRequired,
+  showScoreboard: PropTypes.bool,
+  showGameGuide: PropTypes.bool,
+  gameGuideType: PropTypes.string,
+  gameGuideURL: PropTypes.string,
+  tokenAddress: PropTypes.string.isRequired,
 }
 
 Header.defaultProps = {
@@ -111,6 +149,11 @@ Header.defaultProps = {
   currentBalance: '0',
   currentProvider: {},
   currentAccount: '',
+  isTournament: false,
+  showScoreboard: false,
+  showGameGuide: false,
+  gameGuideType: 'default',
+  gameGuideURL: '',
 }
 
 export default Header
