@@ -1,7 +1,7 @@
 import classNames from 'classnames/bind'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
-import * as css from './index.scss'
+import css from './index.scss'
 
 const cx = classNames.bind(css)
 
@@ -19,7 +19,7 @@ class Block extends PureComponent {
 
     return (
       <div className={cx(margin, className, { center })} style={this.blockStyle} {...props}>
-        { children }
+        {children}
       </div>
     )
   }
