@@ -1,4 +1,4 @@
-import * as walletIntegrations from 'integrations'
+import integrations from 'integrations'
 import { initGnosis } from 'actions/blockchain'
 import { runProviderRegister, runProviderUpdate } from 'integrations/store/actions'
 import { map } from 'lodash'
@@ -15,13 +15,12 @@ export default store => next => (action) => {
       initGnosis: () => dispatch(initGnosis()),
       dispatch,
     }
-
-    Promise.all(map(walletIntegrations, integration => integration.initialize(providerOptions)))
+    Promise.all(map(integrations, integration => integration.initialize(providerOptions)))
   }
 
   if (type === 'PROVIDER_LOGOUT') {
-    Object.keys(walletIntegrations).forEach((providerName) => {
-      const integration = walletIntegrations[providerName]
+    Object.keys(integrations).forEach((providerName) => {
+      const integration = integrations[providerName]
 
       if (integration.constructor.providerName === payload) {
         integration.logout()
