@@ -2,14 +2,16 @@ import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
 import { reducer as formReducer } from 'redux-form'
 import integrations from 'integrations/store/reducers'
+import users from 'routes/Scoreboard/store/reducers/users'
 import market, { REDUCER_ID } from 'store/reducers/market'
+import { isTournament } from 'utils/configuration'
 import entities from './entities'
 import modal from './modal'
 import transactions from './transactions'
 import blockchain from './blockchain'
 import notifications from './notifications'
 
-const reducers = combineReducers({
+const reducers = {
   routing: routerReducer,
   form: formReducer,
   modal,
@@ -19,7 +21,15 @@ const reducers = combineReducers({
   notifications,
   integrations,
   [REDUCER_ID]: market,
-})
+}
+
+if (isTournament()) {
+  reducers.tournament = combineReducers({
+    ranking: users,
+  })
+}
+
+const combinedReducers = combineReducers(reducers)
 
 const rootReducer = (state, action) => {
   let resultState = state
@@ -29,7 +39,7 @@ const rootReducer = (state, action) => {
       ...action.payload,
     }
   }
-  return reducers(resultState, action)
+  return combinedReducers(resultState, action)
 }
 
 export default rootReducer
