@@ -1,26 +1,32 @@
 import { connect } from 'react-redux'
 import { formValueSelector } from 'redux-form'
 import { replace } from 'react-router-redux'
-import { requestGasCost, requestGasPrice } from 'actions/blockchain'
+import { requestGasPrice, requestTokenSymbol } from 'actions/blockchain'
 import MarketDetail from 'routes/MarketDetails/components/MarketDetail'
 
+import { redeemWinnings } from 'actions/market'
 import {
   buyMarketShares,
   sellMarketShares,
-  requestMarketShares,
-  requestMarket,
   requestMarketTrades,
+  requestMarketSharesForAccount,
   requestMarketTradesForAccount,
-  redeemWinnings,
-  withdrawFees,
-} from 'actions/market'
+  requestMarket,
+  requestGasCost,
+} from 'routes/MarketDetails/store/actions'
 import { getMarketById } from 'selectors/market'
-import { getMarketTradesForAccount } from 'selectors/marketTrades'
-import { getMarketShares } from 'selectors/marketShares'
-import { getMarketGraph } from 'selectors/marketGraph'
-import { getGasCosts, getGasPrice, isGasCostFetched, isGasPriceFetched } from 'selectors/blockchain'
+import {
+  getMarketGraph,
+  getMarketTradesForAccount,
+  getMarketShares,
+  getGasCosts,
+  getGasPrice,
+  isGasCostFetched,
+  isGasPriceFetched,
+} from 'routes/MarketDetails/store/selectors'
 import { checkWalletConnection, getCurrentAccount, getCurrentBalance } from 'integrations/store/selectors'
 import { isModerator, getModerators } from 'utils/helpers'
+import { getTokenSymbol } from 'selectors/blockchain'
 
 let marketId
 
@@ -60,12 +66,13 @@ const mapStateToProps = (state, ownProps) => {
     gasCosts: getGasCosts(state),
     gasPrice: getGasPrice(state),
     currentBalance: getCurrentBalance(state),
+    collateralTokenSymbol: getTokenSymbol(state, market.event?.collateralToken),
   }
 }
 
 const mapDispatchToProps = dispatch => ({
   fetchMarket: () => dispatch(requestMarket(marketId)),
-  fetchMarketShares: accountAddress => dispatch(requestMarketShares(marketId, accountAddress)),
+  fetchMarketShares: accountAddress => dispatch(requestMarketSharesForAccount(marketId, accountAddress)),
   fetchMarketTradesForAccount: accountAddress => dispatch(requestMarketTradesForAccount(marketId, accountAddress)),
   fetchMarketTrades: market => dispatch(requestMarketTrades(market)),
   buyShares: (market, outcomeIndex, outcomeTokenCount, cost) =>
@@ -74,9 +81,9 @@ const mapDispatchToProps = dispatch => ({
     dispatch(sellMarketShares(market, outcomeIndex, outcomeTokenCount, earnings)),
   changeUrl: url => dispatch(replace(url)),
   redeemWinnings: market => dispatch(redeemWinnings(market)),
-  withdrawFees: market => dispatch(withdrawFees(market)),
   requestGasCost: (contractType, opts) => dispatch(requestGasCost(contractType, opts)),
   requestGasPrice: () => dispatch(requestGasPrice()),
+  requestTokenSymbol: tokenAddress => dispatch(requestTokenSymbol(tokenAddress)),
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(MarketDetail)
